Header manipulation step has to be skipped if the config is missing
For some reason, error message appears on the form after various actions. Currently i beleive the issue appears
when an email sending is happening in the background. It is hard to track down, because i can't reproduce it on
my environment, so it contains an info log about the context also. My mailing setup might be different.